Fix circular dependency issue with BaseNode
